Add files via upload
This file contains visualisations of the artworks, which are sold, lost or looted from four Hungarian collections (Marcell Nemes, Mór Lipót Herzog, Count Andrássy Gyula, the Younger and Baron Ferenc Hatvany) in the first half of the 20th century. There are five visualisations: the distribution of the 159 items in the four collections, the origin of the artist created the artwork, the number of artworks lost from the four collection, the distribution of the artwork according to owner categories (Private collection, Lost, Museum) and the current location of the artworks, showing only those with more than 2 items.